Files Edited: Inventory.java & fragment_inventory.xml Files Added: PlaceholderFragment.java
Notes:
I edited the Inventory.java to allow the tabs to work correctly. The entire placeholderfragment class inside the Inventory.java I commented out and made a separate class in the project folder called PlaceholderFragment.java. This .java has the same functionality but allows for the tabs to be instanciated individually (or at least that is how I beleave it works).